--- decorators/index.js
@@ -1,5 +1,0 @@
-const controllerWrapper = require('./controllerWrapper');
-const validateBody = require('./validateBody');
-const handleMongooseError = require('./handleMongooseError');
-
-module.exports = { controllerWrapper, validateBody, handleMongooseError };
