+++ helpers/index.js
@@ -1,3 +1,5 @@
 const HttpError = require('./HttpError');
+const controllerWrapper = require('./controllerWrapper');
+const handleMongooseError = require('./handleMongooseError');
 
-module.exports = { HttpError };
+module.exports = { HttpError, controllerWrapper, handleMongooseError };
