+++ middlewares/index.js
@@ -1,9 +1,11 @@
 const globalErrorHandler = require('./globalErrorHandler');
 const isValidId = require('./isValidId');
+const validateBody = require('./validateBody');
 const authenticate = require('./authenticate');
 
 module.exports = {
   globalErrorHandler,
   isValidId,
   authenticate,
+  validateBody,
 };
